rename `use-location.d.ts` to `use-browser-location.d.ts`
